# appcrud/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste Backend Luiz Felipe Tavares das Neves",
        "description": "Todos os endpoins devem ser acompanhados do token de identificação",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API parar cadastro de clientes e pedidos"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Irá autenticar um usuário com base em seu usuario e senha",
                "operationId": "authenticate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@admin.com",
                                    "password": "admin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Será listado json com token de acesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "token": "hash do token"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}